Cast timestamp epoch into FLOAT to avoid rounding
On PostgreSQL-12, the epoch FLOAT is now 6 digits but rounded to 5 digits when
cast into NUMERIC.

issue8779
review264501002
(grafted from 5a45753983a39d6e14a21bd7126a87cd257641c8)

--HG--
branch : 5.2

+++ trytond/model/modelsql.py
@@ -538,7 +538,7 @@ class ModelSQL(ModelStorage):
                         '%s,%s' % (cls.__name__, id_))
                 except KeyError:
                     continue
-                sql_type = fields.Numeric('timestamp').sql_type().base
+                sql_type = fields.Float('timestamp').sql_type().base
                 where.append((table.id == id_)
                     & (Extract('EPOCH',
                             Coalesce(table.write_date, table.create_date)
